+++ src/hash.rs
@@ -109,47 +109,3 @@ pub type Keccak256 = HashFunctionWrapper<sha3::Keccak256>;
 
 // BLAKE2
 pub type Blake2b<DigestLength> = HashFunctionWrapper<blake2::Blake2b<DigestLength>>;
-
-/// XXH3. Note that this is NOT a cryptographic hash function and should NOT be used for cryptographic purposes in production.
-#[cfg(feature = "unsafe_schemes")]
-#[derive(Default)]
-pub struct XXH3 {
-    instance: xxhash_rust::xxh3::Xxh3,
-}
-
-impl OutputSizeUser for XXH3 {
-    type OutputSize = typenum::U8;
-}
-
-impl HashFunction<typenum::U8> for XXH3 {
-    fn update(&mut self, data: &[u8]) {
-        self.instance.update(data);
-    }
-
-    fn finalize(self) -> Digest<typenum::U8> {
-        let hash: [u8; 8] = self.instance.digest().to_be_bytes();
-        Digest(hash.into())
-    }
-}
-
-/// XXH128. Note that this is NOT a cryptographic hash function and should NOT be used for cryptographic purposes in production.
-#[cfg(feature = "unsafe_schemes")]
-#[derive(Default)]
-pub struct XXH128 {
-    instance: xxhash_rust::xxh3::Xxh3,
-}
-
-impl OutputSizeUser for XXH128 {
-    type OutputSize = typenum::U16;
-}
-
-impl HashFunction<typenum::U16> for XXH128 {
-    fn update(&mut self, data: &[u8]) {
-        self.instance.update(data);
-    }
-
-    fn finalize(self) -> Digest<typenum::U16> {
-        let hash: [u8; 16] = self.instance.digest128().to_be_bytes();
-        Digest(hash.into())
-    }
-}
