+++ src/unsecure/signature.rs
@@ -1,14 +1,6 @@
 // Copyright (c) 2022, Mysten Labs, Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-/*
- * This file contains an implementation of a negligible-cost (apart from some ocassional pk copying) trivial
- * cryptographic scheme. The purpose of this library is to allow seamless benchmarking of systems without
- * taking into account the cost of cryptographic primitives - and hence providing a theoretical maximal
- * throughput that a system could achieve if the cost of crypto is optimized away.
- *
- * Warning: All schemes in this file are completely unsafe to use in production.
- */
 use crate::{
     error::FastCryptoError, hash::HashFunction, pubkey_bytes::PublicKeyBytes,
     serde_helpers::keypair_decode_base64,
